Soft Ofertare Usi/requirements.txt
customtkinter
CTkMessagebox
ttkbootstrap
pandas
openpyxl
Pillow
fpdf
requests
supabase
python-dotenv
pyinstaller>=6.0
